Broke out the feature rasterization code to a FeatureRasterizer class from the FeatureImageLayer.
