+++ src/osgEarth/CMakeLists.txt
@@ -334,6 +334,7 @@ SET(LIB_PUBLIC_HEADERS
     FeatureDisplayLayout
     FeatureElevationLayer
     FeatureImageLayer
+    FeatureRasterizer
     FeatureIndex
     TerrainConstraintLayer
     FeatureModelGraph
@@ -441,7 +442,7 @@ SET(LIB_PUBLIC_HEADERS
     FileGDBFeatureSource
 
     SDF
-    
+
     ImGui/ImGui
     ImGui/OsgImGuiHandler
     ImGui/LayersGUI
@@ -714,6 +715,7 @@ set(TARGET_SRC
     FeatureDisplayLayout.cpp
     FeatureElevationLayer.cpp
     FeatureImageLayer.cpp
+    FeatureRasterizer.cpp
     TerrainConstraintLayer.cpp
     FeatureModelGraph.cpp
     FeatureModelLayer.cpp
