+++ src/osgEarth/FeatureImageLayer
@@ -23,55 +23,14 @@
 #include <osgEarth/LayerReference>
 #include <osgEarth/FeatureSource>
 #include <osgEarth/StyleSheet>
+#include <osgEarth/FeatureRasterizer>
 
 namespace osgEarth
 {
-    namespace Util
-    {
-        class OSGEARTH_EXPORT FeatureImageRenderer
-        {
-        public:
-            bool render(
-                const TileKey& key,
-                Session* session,
-                const StyleSheet* styles,
-                osg::Image* target,
-                ProgressCallback* progress) const;
-
-        protected:
-            virtual bool renderFeaturesForStyle(
-                Session* session,
-                const Style& style,
-                const FeatureList& features,
-                const GeoExtent& imageExtent,
-                osg::Image* out_image,
-                Cancelable* progress) const = 0;
-
-            osg::ref_ptr<FeatureFilterChain> _filterChain;
-
-        private:
-            bool queryAndRenderFeaturesForStyle(
-                Session* session,
-                const Style& style,
-                const Query& query,
-                const GeoExtent& imageExtent,
-                osg::Image* out_image,
-                ProgressCallback* progress) const;
-
-            void getFeatures(
-                Session* session,
-                const Query& query,
-                const GeoExtent& imageExtent,
-                FeatureList& features,
-                ProgressCallback* progress) const;
-        };
-    }
-
     /**
      * Rasterizes feature data into an image layer.
      */
-    class OSGEARTH_EXPORT FeatureImageLayer : public osgEarth::ImageLayer,
-                                              public osgEarth::Util::FeatureImageRenderer
+    class OSGEARTH_EXPORT FeatureImageLayer : public osgEarth::ImageLayer
     {
     public: // serialization
         class OSGEARTH_EXPORT Options : public ImageLayer::Options {
@@ -100,22 +59,6 @@ namespace osgEarth
         void setStyleSheet(StyleSheet* styles);
         inline StyleSheet* getStyleSheet() const;
 
-    public: // FeatureImageRenderer
-
-        virtual bool renderFeaturesForStyle(
-            Session* session,
-            const Style& style,
-            const FeatureList& features,
-            const GeoExtent& imageExtent,
-            osg::Image* out_image,
-            Cancelable* progress) const;
-
-
-    public:
-        bool static preProcess(osg::Image* image, bool isSDF);
-
-        bool static postProcess(osg::Image* image, bool isCoverage);
-
     public: // ImageLayer
 
         // Opens the layer and returns a status
@@ -135,6 +78,19 @@ namespace osgEarth
         virtual void init();
 
     protected:
+        osg::Image* render(
+            const TileKey& key,
+            Session* session,
+            const StyleSheet* styles,
+            ProgressCallback* progress
+        ) const;
+
+        void getFeatures(
+            Session* session,
+            const Query& query,
+            const GeoExtent& imageExtent,
+            FeatureList& features,
+            ProgressCallback* progress) const;
 
         virtual ~FeatureImageLayer() { }
 
@@ -142,18 +98,9 @@ namespace osgEarth
         osg::ref_ptr<Session> _session;
         osg::ref_ptr<const FeatureProfile> _featureProfile;
         optional<double> _gamma;
+        osg::ref_ptr<FeatureFilterChain> _filterChain;
 
         void updateSession();
-
-        bool renderFeaturesForStyle(
-            Session*           session,
-            const Style&       style,
-            const FeatureList& features,
-            osg::Referenced*   buildData,
-            const GeoExtent&   imageExtent,
-            osg::Image*        image ) const;
-
-        osg::Image* allocateImage() const;
 
         void establishProfile();
     };
