+++ src/osgEarth/FeatureRasterizer
@@ -1,0 +1,51 @@
+/* -*-c++-*- */
+/* osgEarth - Dynamic map generation toolkit for OpenSceneGraph
+ * Copyright 2020 Pelican Mapping
+ * http://osgearth.org
+ *
+ * osgEarth is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ */
+#ifndef OSGEARTH_FEATURE_RASTERIZER
+#define OSGEARTH_FEATURE_RASTERIZER 1
+
+#include <osgEarth/StyleSheet>
+#include <osgEarth/Session>
+#include <osgEarth/Feature>
+
+namespace osgEarth
+{
+    namespace Util
+    {
+        class OSGEARTH_EXPORT FeatureRasterizer
+        {
+        public:
+            FeatureRasterizer(unsigned int width, unsigned int height, const GeoExtent& extent, const Color& backgroundColor = Color::Transparent);
+
+            void render(
+                Session* session,
+                const Style& style,
+                const FeatureProfile* profile,
+                const FeatureList& in_features) const;
+
+            osg::Image* finalize();
+
+        private:
+            GeoExtent _extent;
+            osg::ref_ptr< osg::Image > _image;
+        };
+    }
+}
+
+
+#endif
